+++ security_monkey/auditor.py
@@ -26,9 +26,10 @@ import datastore
 from security_monkey import app, db
 from security_monkey.watcher import ChangeItem
 from security_monkey.common.jinja import get_jinja_env
-from security_monkey.datastore import User
+from security_monkey.datastore import User, AuditorSettings, Account, ItemAudit, Technology
 from security_monkey.common.utils.utils import send_email
 
+from sqlalchemy import and_
 
 class Auditor(object):
     """
@@ -48,10 +49,11 @@ class Auditor(object):
         self.team_emails = app.config.get('SECURITY_TEAM_EMAIL')
         self.emails = []
         self.emails.extend(self.team_emails)
+
         for account in self.accounts:
-            users = User.query.filter(User.daily_audit_email==True).filter(User.accounts.any(name=accounts[0])).all()
-            new_emails = [user.email for user in users]
-            self.emails.extend(new_emails)
+            users = User.query.filter(User.daily_audit_email==True).filter(User.accounts.any(name=account)).all()
+
+        self.emails.extend([user.email for user in users])
 
     def add_issue(self, score, issue, item, notes=None):
         """
@@ -172,6 +174,7 @@ class Auditor(object):
                     db.session.delete(old_issue)
 
         db.session.commit()
+        self._create_auditor_settings()
 
     def email_report(self, report):
         """
@@ -209,3 +212,56 @@ class Auditor(object):
             return template.render({'items': report_list})
         else:
             return False
+
+    def _create_auditor_settings(self):
+        """
+        Checks to see if an AuditorSettings entry exists for each issue.
+        If it does not, one will be created with disabled set to false.
+        """
+        app.logger.debug("Creating/Assigning Auditor Settings in account {} and tech {}".format(self.accounts, self.index))
+
+        query = ItemAudit.query
+        query = query.join((AuditorSettings, AuditorSettings.id == ItemAudit.auditor_setting_id))
+        query = query.join((Technology, Technology.id == AuditorSettings.tech_id))
+        query = query.filter(Technology.name == self.index)
+        issues = query.filter(ItemAudit.auditor_setting_id == None).all()
+
+        for issue in issues:
+            self._set_auditor_setting_for_issue(issue)
+
+        db.session.commit()
+        app.logger.debug("Done Creating/Assigning Auditor Settings in account {} and tech {}".format(self.accounts, self.index))
+
+    def _set_auditor_setting_for_issue(self, issue):
+
+        auditor_setting = AuditorSettings.query.filter(
+            and_(
+                AuditorSettings.tech_id == issue.item.tech_id,
+                AuditorSettings.account_id == issue.item.account_id,
+                AuditorSettings.issue_text == issue.issue
+            )
+        ).first()
+
+        if auditor_setting:
+            auditor_setting.issues.append(issue)
+            db.session.add(auditor_setting)
+            return auditor_setting
+
+        auditor_setting = AuditorSettings(
+            tech_id=issue.item.tech_id,
+            account_id=issue.item.account_id,
+            disabled=False,
+            issue_text=issue.issue
+        )
+
+        auditor_setting.issues.append(issue)
+        db.session.add(auditor_setting)
+        db.session.commit()
+        db.session.refresh(auditor_setting)
+
+        app.logger.debug("Created AuditorSetting: {} - {} - {}".format(
+            issue.issue,
+            self.index,
+            issue.item.account.name))
+
+        return auditor_setting