+++ security_monkey/views/item_issue.py
@@ -21,8 +21,7 @@ from security_monkey.datastore import Item
 from security_monkey.datastore import Account
 from security_monkey.datastore import Technology
 from security_monkey.datastore import ItemRevision
-from security_monkey import db
-from security_monkey import api
+from security_monkey.datastore import AuditorSettings
 
 from flask.ext.restful import marshal, reqparse
 
@@ -93,6 +92,7 @@ class ItemAuditList(AuthenticatedService):
         self.reqparse.add_argument('names', type=str, default=None, location='args')
         self.reqparse.add_argument('active', type=str, default=None, location='args')
         self.reqparse.add_argument('searchconfig', type=str, default=None, location='args')
+        self.reqparse.add_argument('enabledonly', type=bool, default=None, location='args')
         args = self.reqparse.parse_args()
 
         page = args.pop('page', None)
@@ -128,13 +128,18 @@ class ItemAuditList(AuthenticatedService):
                 (ItemAudit.justification.ilike('%{}%'.format(search))) |
                 (Item.name.ilike('%{}%'.format(search)))
             )
+        if 'enabledonly' in args:
+            query = query.join((AuditorSettings, AuditorSettings.id == ItemAudit.auditor_setting_id))
+            query = query.filter(AuditorSettings.disabled == False)
+
         query = query.order_by(ItemAudit.justified, ItemAudit.score.desc())
         issues = query.paginate(page, count)
 
-        marshaled_dict = {}
-        marshaled_dict['page'] = issues.page
-        marshaled_dict['total'] = issues.total
-        marshaled_dict['auth'] = self.auth_dict
+        marshaled_dict = {
+            'page': issues.page,
+            'total': issues.total,
+            'auth': self.auth_dict
+        }
 
         items_marshaled = []
         for issue in issues.items: